Switch back to long based RLE list implementation

+++ src/main/java/de/uni_koblenz/west/koral/master/statisticsDB/impl/multi_file/storage/ExtraRowStorage.java
@@ -8,7 +8,7 @@ public interface ExtraRowStorage extends RowStorage {
 
 	public void deleteRow(long rowId);
 
-	public byte[] getFreeSpaceIndexData();
+	public long[] getFreeSpaceIndexData();
 
 	public void defragFreeSpaceIndex();
